Move disk cache tests
